Attempt to detect session restore and reload the addon if necessary. The sessions API does not indicate which type of event invoked the onChanged event handler, but as far as I can tell there are 5 options: Closing a tab which has loaded a page other than about:newtab at least once, restoring a closed tab, closing a window, restoring a closed window, and restoring a session. When restoring a session the sessions.onChanged event is invoked only once regardless of the number of windows or tabs restored, and session restore never closes tabs, so tracking the active events and the number of tabs created should allow the detection of session restore.
The restoration is not detected if only a single tab is restored. Differentiating between restoring a single closed tab and restoring a session with a single tab in it likely requires tracking the list of recently closed tabs. Since restoring a session with a single tab in it is unlikely this is not implemented.